fade-in confirm-pickup page elements
