+++ src/components/SignInForm/SignInForm.tsx
@@ -14,7 +14,7 @@ import Link from 'next/link'
 import NextArrow from '@components/SvgComponents/NextArrow'
 import SignUpModule from '@components/SignUpModal'
 import { motion } from 'framer-motion'
-import { container, item } from '@components/SigninModal'
+import { container, item } from '@styles/framer'
 
 const formSchema = z.object({
   email: z.string().email({ message: 'Please enter a valid email' }),
