+++ src/styles/framer.tsx
@@ -1,0 +1,17 @@
+// Framer-motion animation
+
+// Fade-in sequentially
+export const container = {
+  hidden: { opacity: 0 },
+  show: { opacity: 1, transition: { staggerChildren: 0.1 } },
+}
+
+export const item = {
+  hidden: { opacity: 0, translateX: 0, translateY: 0 },
+  show: {
+    opacity: 1,
+    translateX: 0,
+    translateY: 0,
+    transition: { ease: 'easeIn', duration: 0.5 },
+  },
+}
